+++ lib/logic/show_add_weapon_dialog.dart
@@ -90,6 +90,7 @@ Future<void> showAddWeaponDialog({
                                 });
                               }
                             },
+                            allWeapons: weapons,
                           ),
                           isDarkMode: isDarkMode,
                         ),